refactoring celListSlice
